+++ src/main/java/racing/CarRacing.java
@@ -1,5 +1,7 @@
 package racing;
 
+import racing.view.ResultView;
+
 public class CarRacing {
     public static void main(String[] args) {
         new ResultView().play(new Racing().start());
