+++ src/main/java/racing/domain/Grid.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import java.util.LinkedList;
 import java.util.List;
